+++ src/Components/Templates/ContentElements/Accordion.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import {Accordion as ReactAccordion} from "react-bootstrap"
-import Image from "./Image";
-import Gallery from "./Gallery";
+import Gallery from "./Gallery"
 
 const Accordion: React.FC<{ data: any }> = (props) => {
     const {accordionItems, pi_flexform} = props.data
@@ -11,29 +10,27 @@ const Accordion: React.FC<{ data: any }> = (props) => {
         return <></>
     }
 
-    console.log(accordionItems)
-
-
     const accorditionItemsTemplate = accordionItems.map((accordionItem) => {
         let galleryTemplate = <></>
-        if(accordionItem.media.length > 0) {
-            galleryTemplate = <Gallery data={{images: accordionItem.media, ...accordionItem}} />
+        if (accordionItem.media.length > 0) {
+            galleryTemplate = <Gallery data={{images: accordionItem.media, ...accordionItem}}/>
         }
 
-        return <ReactAccordion.Item key={accordionItem.uid} eventKey={accordionItem.uid.toString()} >
+        return <ReactAccordion.Item key={accordionItem.uid} eventKey={accordionItem.uid.toString()}>
             <ReactAccordion.Header as={"h4"} id={`accordion-heading-${accordionItem.uid}`}>
                 <span className="accordion-title-link-text">{accordionItem.header}</span>
             </ReactAccordion.Header>
             <ReactAccordion.Body>
                 <div className={`accordion-content accordion-content-${accordionItem.mediaorient}`}>
                     {galleryTemplate}
-                    <div className={'accordion-content-item accordion-content-text'} dangerouslySetInnerHTML={{__html: accordionItem.bodytext}}/>
+                    <div className={'accordion-content-item accordion-content-text'}
+                         dangerouslySetInnerHTML={{__html: accordionItem.bodytext}}/>
                 </div>
             </ReactAccordion.Body>
         </ReactAccordion.Item>
     })
 
-    return <ReactAccordion defaultActiveKey={activeElement} >
+    return <ReactAccordion defaultActiveKey={activeElement}>
         {accorditionItemsTemplate}
     </ReactAccordion>
 }
